+++ Include/internal/pycore_runtime.h
@@ -23,6 +23,7 @@ extern "C" {
 #include "pycore_pythread.h"        // struct _pythread_runtime_state
 #include "pycore_obmalloc.h"        // struct obmalloc_state
 #include "pycore_time.h"            // struct _time_runtime_state
+#include "pycore_tracemalloc.h"     // struct _tracemalloc_runtime_state
 #include "pycore_unicodeobject.h"   // struct _Py_unicode_runtime_ids
 
 struct _getargs_runtime_state {
@@ -137,11 +138,9 @@ typedef struct pyruntimestate {
     struct _ceval_runtime_state ceval;
     struct _gilstate_runtime_state gilstate;
     struct _getargs_runtime_state getargs;
-    struct {
-        struct _PyTraceMalloc_Config config;
-    } tracemalloc;
     struct _dtoa_runtime_state dtoa;
     struct _fileutils_state fileutils;
+    struct _tracemalloc_runtime_state tracemalloc;
 
     PyPreConfig preconfig;
